Add bindQuery/bindObservableQuery and build the tables on the seam
Add bindQuery and bindObservableQuery — higher-order helpers that pair
Cratis's query + paging behavior (and the paginator) with any
TableRenderer, mirroring the asCommandFormField adapter pattern. The
renderer receives only the UI-library-agnostic TableRendererProps; the
query never leaks across the seam.

Re-express DataTableForQuery as bindQuery(DataTableCore) and
DataTableForObservableQuery as bindObservableQuery(DataTableCore). Public
props (including the PrimeReact pt/ptOptions/unstyled pass-through) and
runtime behavior are unchanged — the PrimeReact-specific props are layered
on at the table wrapper, not in the seam. Export the new surface and
DataTableCore from the DataTables barrel.
