+++ Source/DataTables/DataTableForObservableQuery.tsx
@@ -2,14 +2,24 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 import type { DataTableRootProps } from '@primereact/types/primitive/datatable';
-import { Constructor } from '@cratis/fundamentals';
-import { IObservableQueryFor, Paging } from '@cratis/arc/queries';
-import { useObservableQueryWithPaging } from '@cratis/arc.react/queries';
-import { ReactNode, useState, useRef, useEffect } from 'react';
+import { IObservableQueryFor } from '@cratis/arc/queries';
 import { DataTableCore } from './DataTableCore';
-import { TablePaginator, type TablePaginatorProps } from './TablePaginator';
-import type { DataTableFilterMeta } from './DataTableFilterMeta';
-import type { DataTableSelectionChangeEvent } from './DataTableSelectionChangeEvent';
+import { bindObservableQuery, type BoundObservableQueryTableProps } from './bindObservableQuery';
+
+/**
+ * The PrimeReact styling pass-through the default `DataTableCore` renderer
+ * accepts. Layered onto the observable query binding here (not in the
+ * UI-library-agnostic {@link bindObservableQuery} seam) and forwarded verbatim
+ * to `DataTableCore`.
+ */
+interface DataTablePassThroughProps {
+    /** PrimeReact pass-through configuration applied to the underlying DataTable. */
+    pt?: DataTableRootProps['pt'];
+    /** PrimeReact pass-through options applied to the underlying DataTable. */
+    ptOptions?: DataTableRootProps['ptOptions'];
+    /** When true, disables every base PrimeReact style on the underlying DataTable. */
+    unstyled?: boolean;
+}
 
 /**
  * Props for {@link DataTableForObservableQuery}.
@@ -18,82 +28,8 @@ import type { DataTableSelectionChangeEvent } from './DataTableSelectionChangeEv
  * @typeParam TDataType - The row type returned by the query.
  * @typeParam TArguments - The query's argument object type.
  */
-export interface DataTableForObservableQueryProps<TQuery extends IObservableQueryFor<TDataType, TArguments>, TDataType extends object, TArguments extends object> {
-    /**
-     * Children to render — `<Column>` elements describing the visible columns.
-     */
-    children?: ReactNode;
-
-    /**
-     * The type of query to use
-     */
-    query: Constructor<TQuery>;
-
-    /**
-     * Optional arguments to pass to the query
-     */
-    queryArguments?: TArguments;
-
-    /**
-     * The message to show when there is no data
-     */
-    emptyMessage: string;
-
-    /**
-     * The key to use for the data
-     */
-    dataKey?: string | undefined;
-
-    /**
-     * The current selection.
-     */
-    selection?: TDataType | undefined | null;
-
-    /**
-     * Callback for when the selection changes
-     */
-    onSelectionChange?(event: DataTableSelectionChangeEvent<TDataType>): void;
-
-    /**
-     * Fields to use for global filtering
-     */
-    globalFilterFields?: string[] | undefined;
-
-    /**
-     * Default filters to use
-     */
-    defaultFilters?: DataTableFilterMeta;
-
-    /**
-     * @deprecated No longer toggles behavior. Filtering (`<Column filter>` and the
-     * global search) is always applied client-side to the loaded page; this flag
-     * is retained only for source compatibility and will be removed in a future
-     * release.
-     */
-    clientFiltering?: boolean;
-
-    /**
-     * Extra CSS class name forwarded to the underlying DataTable root.
-     */
-    className?: string;
-
-    /** PrimeReact pass-through configuration applied to the underlying DataTable. */
-    pt?: DataTableRootProps['pt'];
-
-    /** PrimeReact pass-through options applied to the underlying DataTable. */
-    ptOptions?: DataTableRootProps['ptOptions'];
-
-    /** When true, disables every base PrimeReact style on the underlying DataTable. */
-    unstyled?: boolean;
-
-    /** Extra CSS class name forwarded to the paginator. */
-    paginatorClassName?: string;
-
-    /** Accessible names for the paginator controls. Override any to localize. */
-    paginatorAriaLabels?: TablePaginatorProps['ariaLabels'];
-}
-
-const paging = new Paging(0, 20);
+export type DataTableForObservableQueryProps<TQuery extends IObservableQueryFor<TDataType, TArguments>, TDataType extends object, TArguments extends object> =
+    BoundObservableQueryTableProps<TQuery, TDataType, TArguments> & DataTablePassThroughProps;
 
 /**
  * A paged data table bound to a real-time Cratis Arc observable query
@@ -102,6 +38,13 @@ const paging = new Paging(0, 20);
  * underlying read model changes server-side. Rows render through the headless
  * {@link DataTableCore} inside an internally-scrolling region that resizes to
  * fill its container.
+ *
+ * This is `bindObservableQuery(DataTableCore)` — the
+ * {@link bindObservableQuery} observable query/paging behavior paired with the
+ * default `DataTableCore` renderer. To render a query's paged rows with a
+ * *different* table implementation, call `bindObservableQuery` with your own
+ * {@link TableRenderer}. See
+ * [Bring your own table renderer](../../Documentation/DataTables/bring-your-own-renderer.md).
  *
  * ## Children
  *
@@ -118,7 +61,7 @@ const paging = new Paging(0, 20);
  * ```
  *
  * Use {@link DataTableForQuery} for one-shot snapshot queries. Use
- * {@link DataPage} for a higher-level layout that combines this table with
+ * `DataPage` for a higher-level layout that combines this table with
  * an action menubar, selection, and a details pane.
  *
  * @typeParam TQuery - The query class (proxy generated from C# `IObservableQueryFor`).
@@ -126,98 +69,4 @@ const paging = new Paging(0, 20);
  * @typeParam TArguments - The query's argument object type.
  * @param props - {@link DataTableForObservableQueryProps}.
  */
-export const DataTableForObservableQuery = <TQuery extends IObservableQueryFor<TDataType, TArguments>, TDataType extends object, TArguments extends object>(props: DataTableForObservableQueryProps<TQuery, TDataType, TArguments>) => {
-    const [result, , setPage] = useObservableQueryWithPaging(props.query, paging, props.queryArguments);
-    const containerRef = useRef<HTMLDivElement>(null);
-    const [tableHeight, setTableHeight] = useState<number>(600);
-    const timeoutRef = useRef<ReturnType<typeof setTimeout> | undefined>(undefined);
-    const totalItems = result.paging.totalItems;
-    const pageCount = result.paging.totalPages;
-    const showPaginator = totalItems > 0 && pageCount > 1;
-
-    useEffect(() => {
-        if (!containerRef.current) return;
-
-        const resizeObserver = new ResizeObserver((entries) => {
-            if (timeoutRef.current) {
-                clearTimeout(timeoutRef.current);
-            }
-
-            timeoutRef.current = setTimeout(() => {
-                for (const entry of entries) {
-                    const containerHeight = entry.contentRect.height;
-                    if (containerHeight > 0) {
-                        const paginatorHeight = showPaginator ? 56 : 0;
-                        const calculatedHeight = containerHeight - paginatorHeight - 2;
-                        const newHeight = Math.max(calculatedHeight, 200);
-
-                        setTableHeight(prevHeight => {
-                            if (Math.abs(newHeight - prevHeight) > 5) {
-                                return newHeight;
-                            }
-                            return prevHeight;
-                        });
-                    }
-                }
-            }, 10);
-        });
-
-        resizeObserver.observe(containerRef.current);
-
-        return () => {
-            if (timeoutRef.current) {
-                clearTimeout(timeoutRef.current);
-            }
-            resizeObserver.disconnect();
-        };
-    }, [showPaginator]);
-
-    return (
-        <div
-            ref={containerRef}
-            style={{
-                display: 'flex',
-                flexDirection: 'column',
-                height: '100%',
-                border: '1px solid var(--cratis-surface-border)',
-                borderRadius: 'var(--cratis-border-radius)',
-                overflow: 'hidden'
-            }}>
-            <div style={{ height: `${tableHeight}px`, overflow: 'hidden' }}>
-                <DataTableCore<TDataType>
-                    data={result.data as unknown as TDataType[]}
-                    dataKey={props.dataKey}
-                    emptyMessage={props.emptyMessage}
-                    selectionMode='single'
-                    selection={props.selection}
-                    onSelectionChange={props.onSelectionChange}
-                    globalFilterFields={props.globalFilterFields}
-                    defaultFilters={props.defaultFilters}
-                    scrollable
-                    scrollHeight='100%'
-                    className={props.className}
-                    style={{ minWidth: '100%' }}
-                    pt={props.pt}
-                    ptOptions={props.ptOptions}
-                    unstyled={props.unstyled}>
-                    {props.children}
-                </DataTableCore>
-            </div>
-
-            {showPaginator && (
-                <div style={{ borderTop: '1px solid var(--cratis-surface-border)', flexShrink: 0 }}>
-                    <TablePaginator
-                        page={result.paging.page}
-                        pageCount={pageCount}
-                        onPageChange={setPage}
-                        totalItems={totalItems}
-                        pageSize={paging.pageSize}
-                        className={props.paginatorClassName}
-
-                        ariaLabels={props.paginatorAriaLabels}
-                    />
-                </div>
-            )}
-        </div>
-    );
-};
+export const DataTableForObservableQuery = bindObservableQuery<DataTablePassThroughProps>(DataTableCore);
